// src/contracts/ui-types.ts
export type FieldMapping = {
    key: string;
    description: string;
    type?: string;
};

export const FIELD_MAPPINGS: Record<string, Record<string, FieldMapping>> = {
    item: {
        name: {
            key: 'item.name',
            description: 'Item Name',
        },
        externalReference: {
            key: 'item',
            description: 'Item External Reference',
        },
    },
    productVariant: {
        name: {
            key: 'variant.name',
            description: 'Variant Name',
        },
        sku: {
            key: 'variant.sku',
            description: 'Variant SKU',
        },
        images: {
            key: 'variant.images',
            description: 'Variant Images',
        },
        price: {
            key: 'variant.price',
            description: 'Variant Price',
        },
        stock: {
            key: 'variant.stock',
            description: 'Variant Stock',
        },
        externalReference: {
            key: 'variant.externalReference',
            description: 'Variant External Reference',
        },
    },
};